LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-audioworklet-interface/audioworklet-audioparam-size.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: undefined is not an object (evaluating 'nodeParam.setValueAtTime')

Harness Error (FAIL), message = Unhandled rejection: undefined is not an object (evaluating 'nodeParam.setValueAtTime')

PASS # AUDIT TASK RUNNER STARTED.
PASS Executing "Initializing AudioWorklet and Context"
TIMEOUT Executing "Verify Size of AudioParam Arrays" Test timed out
NOTRUN Audit report
PASS > [Initializing AudioWorklet and Context]
PASS   Creating offline context for testing did not throw an exception.
PASS   Creating test worklet resolved correctly.
PASS < [Initializing AudioWorklet and Context] All assertions passed. (total 2 assertions)
PASS > [Verify Size of AudioParam Arrays]
